Implement Color mix (#19)
* add mix

* Add Add() & Mul<f64>() for Color

* 8 implementation du raycasting (#20)

* Algorithme de generation fait et non tester(Requiert changement a vec3)

* Fixed conflicts

* Imbriquement des differentes pieces apres rebase

* Fix de la generation des rays + Intersections

---------


* Graphics library change (#21)

* display vector of vector

* sped up display

* 8 implementation du raycasting (#20)

* Algorithme de generation fait et non tester(Requiert changement a vec3)

* Fixed conflicts

* Imbriquement des differentes pieces apres rebase

* Fix de la generation des rays + Intersections

---------


* Changed graphics library from nannou to pixels, image and winit.
Merged Johrober's code
Benchmark printing

* small fixes

---------


* add mix

* Add Add() & Mul<f64>() for Color

* Change colors to float

---------

+++ src/model/materials/unicolor.rs
@@ -9,7 +9,7 @@ pub struct Unicolor {
 }
 
 impl Unicolor {
-    pub fn new(r: u8, g: u8, b: u8)-> Self {
+    pub fn new(r: f64, g: f64, b: f64)-> Self {
         Self {
             color: Color::new(r, g, b)
         }
